refactor ♻️: ネットワーク機能のコア実装への移行とファイル構造整理
