+++ src/core/nic_worker.rs
