+++ src/core/ping_worker.rs
