--- src/network.rs
@@ -1,6 +1,0 @@
-pub mod nic_worker;
-pub mod ping_worker;
-pub mod traceroute_worker;
-pub mod worker_pool;
-
-pub use worker_pool::WorkerPool;
